# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Magic Garden API
  # Version du *contrat*, pas du serveur ni du jeu. Un client la compare à celle
  # qu'il sait lire avant de faire confiance à une URL. Chaîne parce que
  # l'OpenAPI 3.0 l'exige ; /schema.json et /data/version la servent en nombre.
  version: "1"
  description: |
    **Unofficial API** for Magic Garden game data that dynamically extracts and provides game content.

    ## Overview

    This API provides two types of data:
    - **Static data**: Extracted from the game's minified bundle (plants, pets, items, etc.)
    - **Live data**: Shops and weather, tracked from the game's official platform API

    ## Quick Start

    The examples below are relative on purpose. This document describes the instance
    that served it, and the base URL is the first entry of `servers` -- a fork is
    deployed under its own name, so an example that named one deployment would send
    every reader to that one.

    ### Get All Game Data
    ```bash
    curl "$API/data"
    ```

    ### Get Specific Plant
    ```bash
    curl "$API/data/plants" | jq '.Carrot'
    ```

    ### Stream Live Updates
    ```javascript
    const eventSource = new EventSource(`${API}/live/stream`);
    eventSource.addEventListener('weather', (event) => {
      const data = JSON.parse(event.data);
      console.log('Weather:', data.weather);
    });
    eventSource.addEventListener('shops', (event) => {
      const shops = JSON.parse(event.data);
...
    });
    ```

...

...

...

...

...

...

...
paths:
...
  /assets/proxy:
    get:
      tags:
        - Assets
      summary: Fetch an upstream asset through this host, with CORS
      description: |
        Streams a whitelisted upstream asset — a `https://magicgarden.gg/…` URL, or
        a `/assets/sprites/…` URL of the upstream fork's sprite host — back to the
        caller with `Access-Control-Allow-Origin: *`. Upstream sets no such header,
        so a browser cannot read those pixels directly; this proxy is the path that
        makes them readable cross-origin.

        **Not a general open proxy:** a URL outside the whitelist is rejected with
        `BAD_REQUEST` and never fetched. The response is cached
        `public, max-age=86400, immutable`, because the URL names a versioned asset.
        An upstream that accepts the connection but never answers becomes a named
        `504 UPSTREAM_TIMEOUT` instead of an open socket.
      parameters:
        - name: url
          in: query
          required: true
          description: Absolute upstream URL to fetch. Only whitelisted hosts are accepted.
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: The upstream body, with the upstream `Content-Type`
          content:
            image/png:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Missing `url`, or a URL outside the whitelist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Upstream answered a non-2xx status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: Upstream did not respond within the platform timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /assets/sprite-data:
    get:
      tags:
        - Assets
      summary: Get sprites data
      description: |
        Returns sprite atlas metadata with powerful filtering options.

        **Query parameters:**
        - `search`: Filter sprites by name (case-insensitive partial match)
        - `cat`: Filter by category (seeds, plants, pets, etc.)
        - `full=1`: Include detailed sprite information (anchors, frames, source sizes)
        - `flat=1`: Return flat array instead of grouped by category

        **Use cases:**
        - Find sprite URLs for specific items
        - Browse available sprites by category
        - Get sprite dimensions and atlas coordinates
        - Build sprite catalogs

        **Response notes:**
        - `baseUrl` points to the game's versioned asset base, not the API host.
        - `url` points to the atlas PNG; use `/assets/sprites` for individual sprite PNGs.

        **Examples:**
        - `/assets/sprite-data?search=Carrot` - Find all Carrot sprites
        - `/assets/sprite-data?cat=seeds` - Get all seed sprites
        - `/assets/sprite-data?search=Bunny&full=1` - Get detailed Bunny sprite info
        - `/assets/sprite-data?cat=pets&flat=1` - Flat list of pet sprites
      parameters:
        - name: full
          in: query
          description: Include full sprite information (anchors, source sizes)
          schema:
            type: string
            enum: ['1']
          example: '1'
        - name: search
          in: query
          description: Filter sprites by ID (partial match, case-insensitive)
          schema:
            type: string
          example: 'Carrot'
        - name: cat
          in: query
          description: Filter by category
          schema:
            type: string
            enum: ["seeds", "plants", "tallPlants", "mutations", "pets", "pet-cosmetic", "decor", "items", "objects", "ui", "animations", "weather", "tiles", "winter"]
          example: 'seeds'
        - name: flat
          in: query
          description: Return flat list instead of grouped by category
          schema:
            type: string
            enum: ['1']
          example: '1'
      responses:
        '200':
          description: Sprites data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpritesResponse'
              examples:
                grouped:
                  summary: Grouped response (default)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 3
                    categories:
                      - cat: "seeds"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                          - id: "Cabbage"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - cat: "plants"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                flat:
                  summary: Flat response (with flat=1)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 3
                    items:
                      - id: "Carrot"
                        cat: "seeds"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - id: "Cabbage"
                        cat: "seeds"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                      - id: "Carrot"
                        cat: "plants"
                        url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                fullDetails:
                  summary: Full details (with full=1)
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 1
                    categories:
                      - cat: "seeds"
                        items:
                          - id: "Carrot"
                            url: "https://magicgarden.gg/version/<version>/assets/sprites-0.png"
                            frame:
                              x: 0
                              y: 0
                              w: 64
                              h: 64
                            anchor:
                              x: 0.5
                              y: 0.5
                            sourceSize:
                              w: 64
                              h: 64
  /assets/cosmetics:
    get:
      tags:
        - Assets
      summary: Get cosmetics data
      description: |
        Returns all cosmetic items for player avatar customization.

        **Categories include:**
        - Hats
        - Outfits
        - Accessories
        - Hair styles
        - Face customization

        **Query parameters:**
        - `full=1`: Include detailed sprite information (frames, anchors)

        **Use cases:**
        - Build character customization UI
        - Display cosmetic shop
        - Show available customization options

        **Response notes:**
        - `baseUrl` points to the game's versioned asset base, not the API host.
      parameters:
        - name: full
          in: query
          description: Include full cosmetic information (sprite frames, anchors)
          schema:
            type: string
            enum: ['1']
          example: '1'
      responses:
        '200':
          description: Cosmetics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CosmeticsResponse'
              examples:
                cosmeticsData:
                  summary: Example cosmetics
                  value:
                    baseUrl: "https://magicgarden.gg/version/<version>/assets/"
                    count: 25
                    categories:
                      - cat: "hats"
                        items:
                          - id: "StrawHat"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/hats/StrawHat.png"
                          - id: "WizardHat"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/hats/WizardHat.png"
                      - cat: "outfits"
                        items:
                          - id: "FarmerOutfit"
                            url: "https://magicgarden.gg/version/<version>/assets/cosmetic/outfits/FarmerOutfit.png"
...
  /assets/sprites:
    get:
      tags:
        - Assets
      summary: Full sprite catalog with ready-to-use PNG URLs
      description: |
        Returns the complete catalog of exported sprites, grouped by category,
        with a ready-to-use PNG URL for each one. Around 600 sprites in total.

        Want a subset? Filter with `cat` (one category) and/or `search`
        (case-insensitive match on the sprite name).

        **Available categories:**
        - `seeds` - Seed sprites
        - `plants` - Plant growth stages
        - `tallPlants` - Tall plant varieties
        - `mutations` - Mutation variants
        - `pets` - Pet companions
        - `pet-cosmetic` - Hats and accessories pets can wear
        - `decor` - Garden decorations
        - `items` - Tools and consumables
        - `objects` - Interactive objects
        - `ui` - UI elements
        - `animations` - Animation frames
        - `weather` - Weather effects
        - `tiles` - Ground tiles
        - `winter` - Winter seasonal sprites

        **Use cases:**
        - Discover every sprite the game currently ships
        - Build sprite pickers or galleries without hardcoding names
        - Find a sprite's exact filename before calling `/assets/sprites/{category}/{name}`

        **Examples:**
        - `/assets/sprites` - everything
        - `/assets/sprites?cat=pets` - all pet sprites
        - `/assets/sprites?search=carrot` - anything named like "carrot", across categories

        **Response notes:**
        - `baseUrl` reflects the public API base used for sprite URLs (`SPRITES_BASE_URL`).
        - URLs include a `?v=<gameVersion>` cache-busting parameter.
        - `categories` always lists every valid category, even when you filter.
      parameters:
        - name: cat
          in: query
          required: false
          schema:
            type: string
            enum: [seeds, plants, tallPlants, mutations, pets, pet-cosmetic, decor, items, objects, ui, animations, weather, tiles, winter]
          description: Only return sprites from this category
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on sprite names
          example: "carrot"
      responses:
        '200':
          description: Sprite catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Number of sprites in the response (after filters)
                    example: 600
                  baseUrl:
                    type: string
                    description: Base URL for the API
                    example: ""
                  categories:
                    type: array
                    items:
                      type: string
                    description: Every valid category (not affected by filters)
                    example: ["animations", "decor", "items", "mutations", "objects", "pets", "plants", "seeds", "tallPlants", "tiles", "ui", "weather", "winter"]
                  sprites:
                    type: object
                    description: Sprites grouped by category, each with a ready-to-use PNG URL
                    additionalProperties:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                            example: "Carrot"
                          url:
                            type: string
                            example: "/assets/sprites/seeds/Carrot.png?v=642"
              examples:
                filtered:
                  summary: Filtered catalog (?search=carrot)
                  value:
                    count: 3
                    baseUrl: ""
                    categories: ["animations", "decor", "items", "mutations", "objects", "pets", "plants", "seeds", "tallPlants", "tiles", "ui", "weather", "winter"]
                    sprites:
                      plants:
                        - name: "BabyCarrot"
                          url: "/assets/sprites/plants/BabyCarrot.png?v=642"
                        - name: "Carrot"
                          url: "/assets/sprites/plants/Carrot.png?v=642"
                      seeds:
                        - name: "Carrot"
                          url: "/assets/sprites/seeds/Carrot.png?v=642"
        '400':
          description: Unknown category in `cat`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /assets/animations:
    get:
      tags:
        - Assets
      summary: Catalog of animated pet loops (WebP/GIF)
      description: |
        Everything the game animates is vector art, so beyond the still PNGs
        this API serves those subjects **animated**: one looping WebP (or GIF)
        per subject and per state.

        Two categories:
        - `pets` - one loop per species and per state (`idle`, `walk`, `eat`,
          `sleep`), plus the two weather variants.
        - `decor` - the eight animated garden decorations (windmill, fountain,
          cauldron, …). Each has a single continuous loop, keyed `loop`.

        Loops are pre-rendered whenever the game ships a new pet file - never
        on request - and served as plain files. Drop the URL in an `<img>` tag
        and it plays; there is no runtime to embed.

        **Clips:**
        - `idle` - the resting loop, 7 s
        - `walk` - one step cycle, ~0.6 s
        - `eat` - eating, 1 s
        - `sleep` - sleeping, 8 s

        Not every species exposes every clip, and the two weather variants
        (`FireHorseActive`, `ThunderWolfActive`) only ship `idle`. Decor clip
        names are taken from what each artboard declares, never hardcoded - the
        game's own timeline names are inconsistent.

        **Filters:** `cat`, `name` (exact species), `clip`, `search`
        (case-insensitive substring on the species name).

        **Response notes:**
        - Every loop is rendered to the same subject height, so species stay
          visually comparable - a snail and a horse come out the same size.
        - `url` is the link, in the primary format (WebP by default); `format`
          names it, and any additional format appears under its own key
          (`gif`).
        - `anchor` places the sprite like the game does: `x` is always 0.5,
          `y` marks the ground line under the pet's feet.
        - `source.generatedAt` tells you when the loops were last rendered, and
          `source.riveUrl` which game file they came from.
        - Not to be confused with the `animations` **sprite** category
          (`/assets/sprites/animations/*.png`), which holds still frames from
          the game's atlases.
      parameters:
        - name: cat
          in: query
          required: false
          schema:
            type: string
            enum: [pets, decor]
          description: Only return animations from this category
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Exact species name (case-insensitive)
          example: "Chicken"
        - name: clip
          in: query
          required: false
          schema:
            type: string
            enum: [idle, walk, eat, sleep, loop]
          description: Only return this clip (`loop` for decor)
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on species names
          example: "horse"
      responses:
        '200':
          description: Animation catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 118
                  baseUrl:
                    type: string
                    example: ""
                  categories:
                    type: array
                    items:
                      type: string
                    example: ["pets", "decor"]
                  formats:
                    type: array
                    items:
                      type: string
                    description: Formats currently generated
                    example: ["webp"]
                  sources:
                    type: object
                    description: |
                      One entry per category - each comes from its own Rive file,
                      and their exports do not finish at the same time.
                    additionalProperties:
                      type: object
                      properties:
                        riveUrl:
                          type: string
                          description: Game file the loops were rendered from
                        generatedAt:
                          type: string
                          format: date-time
                  animations:
                    type: object
                    description: Animations grouped by category
                    additionalProperties:
                      type: array
                      items:
                        $ref: '#/components/schemas/PetAnimation'
              examples:
                filtered:
                  summary: One species (?name=Chicken)
                  value:
                    count: 4
                    baseUrl: ""
                    categories: ["pets"]
                    formats: ["webp"]
                    sources:
                      pets:
                        riveUrl: "https://magicgarden.gg/runtime-assets/pets.bfec8b856f6e2b0e4a83.riv"
                        generatedAt: "2026-08-04T18:21:07.633Z"
                    animations:
                      pets:
                        - category: "pets"
                          name: "Chicken"
                          clip: "idle"
                          timeline: "Pet_Idle"
                          url: "/assets/animations/pets/Chicken_idle.webp?v=824"
                          format: "webp"
                          width: 212
                          height: 257
                          frames: 210
                          fps: 30
                          durationMs: 7000
                          anchor:
                            x: 0.5
                            y: 1
        '400':
          description: Unknown category in `cat`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /assets/rive:
    get:
      tags:
        - Assets
      summary: The game's Rive (vector) files and what they contain
      description: |
        Everything that moves in the game is vector art, rendered live by a Rive
        runtime: the pets, but also the animated decor, the thought bubble, the
        coin and the gift box. This endpoint publishes the files themselves and
        an inventory of what is inside them.

        **Why you may want the file rather than the images.** `pets.riv` is
        3 MB and holds all 28 species with about 660 timelines, at any
        resolution and frame rate. The equivalent coverage as pre-encoded WebP
        would run to hundreds of megabytes. If your client can run the Rive
        runtime, play the source; if it can only show a picture, use
        `/assets/animations`.

        **What the inventory is for.** Artboard names, state machine names,
        timeline names and - above all - **the type of every state machine
        input**. Half of a pet's inputs are *triggers* (`walk`, `eat`,
        `petted`), not booleans (`sleep`, `fire`, `thunder`), and asking for the
        wrong one fails silently. All of it changes with game updates, so
        nothing here should ever be hardcoded by a client.

        **Notes:**
        - `url` goes through `/assets/proxy`, which adds the CORS headers
          magicgarden.gg does not send - a browser cannot fetch `origin`
          directly.
        - `origin` is versioned by content hash, so it changes on every update.
        - `loadable: false` means our pinned runtime cannot read the file.
          `avatar.riv` is permanently in this state: it references assets stored
          outside the file (the cosmetics bundle) and never finishes loading.
        - The inventory is refreshed during sprite sync, never per request:
          inspecting a file costs a download and a WASM pass.
      parameters:
        - name: key
          in: query
          required: false
          schema:
            type: string
          description: Return a single file, with full artboard detail
          example: "pets"
        - name: full
          in: query
          required: false
          schema:
            type: string
            enum: ["1"]
          description: Include every artboard's timelines and state machines
      responses:
        '200':
          description: Rive inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 6
                  baseUrl:
                    type: string
                    example: ""
                  generatedAt:
                    type: string
                    format: date-time
                    description: When the inventory was last refreshed
                  files:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/RiveFile'
              examples:
                summary:
                  summary: Default view
                  value:
                    count: 6
                    baseUrl: ""
                    generatedAt: "2026-08-04T18:25:17.004Z"
                    files:
                      pets:
                        key: "pets"
                        aliases: ["rive/pets.riv", "rive/pets"]
                        bundle: "default"
                        url: "/assets/proxy?url=https%3A%2F%2Fmagicgarden.gg%2Fruntime-assets%2Fpets.bfec8b856f6e2b0e4a83.riv"
                        origin: "https://magicgarden.gg/runtime-assets/pets.bfec8b856f6e2b0e4a83.riv"
                        bytes: 3085491
                        loadable: true
                        artboardCount: 29
                        timelineCount: 638
                        artboards: ["Pets", "Pig", "Horse", "Chicken"]
                      avatar:
                        key: "avatar"
                        bundle: "default"
                        loadable: false
                        error: "Rive load timed out after 20000ms"
        '404':
          description: Unknown `key`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    SpritesResponse:
      type: object
      properties:
        baseUrl:
          type: string
          description: Game assets base URL (versioned)
        count:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              cat:
                type: string
              items:
                type: array
                items:
                  type: object
    CosmeticsResponse:
      type: object
      properties:
        baseUrl:
          type: string
          description: Game assets base URL (versioned)
        count:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              cat:
                type: string
              items:
                type: array
                items:
                  type: object
...
    Error:
      type: object
      description: Standard error response format
      properties:
        error:
          type: object
          description: Error details
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "SPRITE_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "The requested sprite was not found"
            details:
              type: object
              description: Additional error context (varies by error type)
              additionalProperties: true
              example:
                category: "seeds"
                filename: "NonExistent.png"
      example:
        error:
          code: "SPRITE_NOT_FOUND"
          message: "The requested sprite was not found"
          details:
            category: "seeds"
            filename: "NonExistent.png"

    # Le bloc de provenance des corps de `/data/*`. La clé est réservée : les
    # corps sont les données du jeu keyées par nom, et un nom peut être
    # n'importe quoi. Si une catégorie porte déjà `_meta`, le bloc recule
    # (`__meta`) plutôt que d'écraser une entité.
...